WA-376 Moved secure store
