+++ Ethereum/EthereumImplementations/DBExternallyOwnedAccountRepository.swift
@@ -1,0 +1,18 @@
+//
+//  Copyright © 2018 Gnosis Ltd. All rights reserved.
+//
+
+import Foundation
+import EthereumDomainModel
+import Database
+
+public class DBExternallyOwnedAccountRepository: ExternallyOwnedAccountRepository {
+
+    private let db: Database
+
+    public init (db: Database) {
+        self.db = db
+    }
+
+}
+
